# feat: implement performance benchmarking framework