--- a/.github/workflows/performance.yml
+++ b/.github/workflows/performance.yml
@@ -1,8 +1,6 @@
 name: Performance Tests
 
 on:
-  pull_request:
-    branches: [ main ]
   push:
     branches: [ main ]
   workflow_dispatch: