## Summary

Implements a comprehensive performance benchmarking framework for Union Square proxy to ensure MVP performance targets are met.

Closes #11

## Changes

- **Benchmark validation tests** - Tests that verify our benchmark framework can accurately measure sub-millisecond operations and validate <5ms latency requirements
- **Memory profiling with dhat** - Set up memory allocation tracking to identify allocation hotspots and optimize memory usage
- **Load testing scenarios** - Comprehensive load tests for MVP targets:
  - 500 RPS sustained load (30 seconds)
  - 2000 RPS burst load (10 seconds)  
  - 1000 concurrent users (20 seconds)
- **CI/CD integration** - Added performance workflow that:
  - Runs on every PR to detect regressions
  - Compares benchmarks against base branch
  - Posts detailed performance reports as PR comments
  - Fails if regressions exceed 10% threshold
- **Documentation** - Updated benches/README.md with MVP performance targets and load testing instructions

## Test Plan

- [x] Benchmark validation tests pass
- [x] Memory profiling runs successfully
- [x] Load tests are properly marked as ignored (require significant resources)
- [x] CI workflow properly configured
- [x] All existing tests still pass

## Notes

- Load tests are marked as `#[ignore]` since they require proper configuration and significant resources to run
- Performance CI workflow will automatically detect regressions on future PRs
- Memory profiling can be run with: `cargo bench --bench memory_profiling`